Moved codegen out of session.py
